TestFrontmanHubInput: reduce execution time to 200 ms, add stats checks

+++ config.go
@@ -80,7 +80,7 @@ type Config struct {
 
 	SenderBatchSize int `toml:"sender_batch_size" comment:"Do not send back more than N results per POST request"`
 
-	SenderInterval int `toml:"sender_interval" comment:"Make a pause of N seconds between POST requests when processing the result queue"`
+	SenderInterval float64 `toml:"sender_interval" comment:"Make a pause of N seconds between POST requests when processing the result queue"`
 
 	HealthChecks HealthCheckConfig `toml:"health_checks" comment:"Frontman can verify a reliable internet uplink by pinging some reference hosts before each check round starts.\nPing all hosts of the list.\nOnly if frontman gets a positive answer form all of them, frontman continues.\nOtherwise, the entire check round is skipped. No data is sent back.\nFailed health checks are recorded to the log.\nOnly 0% packet loss is considered as a positive check result. Pings are performed in parallel.\nDisabled by default. Enable by declaring reference_ping_hosts targets\n"`
